RavenDB-19689 properly wait for indexes to be deployed to all databases after smuggler import
