+++ src/Raven.Server/Smuggler/Documents/ShardedDatabaseSmuggler.cs
@@ -2,20 +2,13 @@
 using System.Threading;
 using System.Threading.Tasks;
 using Raven.Client;
-using Raven.Client.Documents.Indexes;
 using Raven.Client.Documents.Operations;
 using Raven.Client.Documents.Smuggler;
 using Raven.Client.ServerWide;
-using Raven.Client.Util;
-using Raven.Server.Config;
 using Raven.Server.Documents;
-using Raven.Server.Documents.Indexes;
-using Raven.Server.Documents.Indexes.Auto;
-using Raven.Server.Documents.Indexes.MapReduce.Auto;
 using Raven.Server.Documents.PeriodicBackup;
 using Raven.Server.ServerWide;
 using Raven.Server.ServerWide.Commands;
-using Raven.Server.ServerWide.Commands.Indexes;
 using Raven.Server.Smuggler.Documents.Data;
 using Raven.Server.Smuggler.Documents.Processors;
 using Sparrow.Json;
@@ -53,154 +46,6 @@ namespace Raven.Server.Smuggler.Documents
                              LoggingSource.Instance.GetLogger<DatabaseDestination>(_databaseRecord.DatabaseName)))
             {
                 return await ProcessDatabaseRecordInternalAsync(result, action);
-            }
-        }
-
-        protected override async Task<SmugglerProgressBase.Counts> ProcessIndexesAsync(SmugglerResult result)
-        {
-            result.Indexes.Start();
-
-            var configuration = DatabasesLandlord.CreateDatabaseConfiguration(_server.Server.ServerStore, _databaseRecord.DatabaseName, _databaseRecord.Settings);
-
-            await foreach (var index in _source.GetIndexesAsync())
-            {
-                _token.ThrowIfCancellationRequested();
-                result.Indexes.ReadCount++;
-
-                if (index == null)
-                {
-                    result.Indexes.ErroredCount++;
-                    continue;
-                }
-
-                if (OnIndexAction != null)
-                {
-                    OnIndexAction(index);
-                    continue;
-                }
-
-                switch (index.Type)
-                {
-                    case IndexType.AutoMap:
-                        await PutAutoMapIndex(result, index, configuration);
-                        break;
-
-                    case IndexType.AutoMapReduce:
-                        await PutAutoMapReduceIndex(result, index, configuration);
-                        break;
-
-                    case IndexType.Map:
-                    case IndexType.MapReduce:
-                    case IndexType.JavaScriptMap:
-                    case IndexType.JavaScriptMapReduce:
-                        await PutIndex(result, index, configuration);
-                        break;
-
-                    case IndexType.Faulty:
-                        break;
-
-                    default:
-                        throw new NotSupportedException(index.Type.ToString());
-                }
-
-                if (result.Indexes.ReadCount % 10 == 0)
-                {
-                    var message = $"Read {result.Indexes.ReadCount:#,#;;0} indexes.";
-                    AddInfoToSmugglerResult(result, message);
-                }
-            }
-            return result.Indexes;
-        }
-
-        private async Task PutAutoMapReduceIndex(SmugglerResult result, IndexDefinitionAndType index, RavenConfiguration configuration)
-        {
-            var autoMapReduceIndexDefinition = (AutoMapReduceIndexDefinition)index.IndexDefinition;
-            try
-            {
-                if (IndexStore.IsValidIndexName(autoMapReduceIndexDefinition.Name, false, out var errorMessage) == false)
-                {
-                    throw new ArgumentException(errorMessage);
-                }
-
-                autoMapReduceIndexDefinition.DeploymentMode = configuration.Indexing.AutoIndexDeploymentMode;
-
-                var command = PutAutoIndexCommand.Create(autoMapReduceIndexDefinition, _databaseRecord.DatabaseName, RaftIdGenerator.DontCareId,
-                    configuration.Indexing.AutoIndexDeploymentMode);
-
-                await _server.SendToLeaderAsync(command).ConfigureAwait(false);
-            }
-            catch (Exception e)
-            {
-                result.Indexes.ErroredCount++;
-                result.AddError($"Could not write auto map-reduce index '{autoMapReduceIndexDefinition.Name}': {e.Message}");
-            }
-        }
-
-        private async Task PutAutoMapIndex(SmugglerResult result, IndexDefinitionAndType index, RavenConfiguration configuration)
-        {
-            var autoMapIndexDefinition = (AutoMapIndexDefinition)index.IndexDefinition;
-
-            try
-            {
-                if (IndexStore.IsValidIndexName(autoMapIndexDefinition.Name, false, out var errorMessage) == false)
-                {
-                    throw new ArgumentException(errorMessage);
-                }
-
-                autoMapIndexDefinition.DeploymentMode = configuration.Indexing.AutoIndexDeploymentMode;
-
-                CommandBase command = PutAutoIndexCommand.Create(autoMapIndexDefinition, _databaseRecord.DatabaseName, RaftIdGenerator.DontCareId,
-                    configuration.Indexing.AutoIndexDeploymentMode);
-
-                await _server.SendToLeaderAsync(command).ConfigureAwait(false);
-            }
-            catch (Exception e)
-            {
-                result.Indexes.ErroredCount++;
-                result.AddError($"Could not write auto map index '{autoMapIndexDefinition.Name}': {e.Message}");
-            }
-        }
-
-        private async Task PutIndex(SmugglerResult result, IndexDefinitionAndType index, RavenConfiguration configuration)
-        {
-            var indexDefinition = (IndexDefinition)index.IndexDefinition;
-            if (string.Equals(indexDefinition.Name, "Raven/DocumentsByEntityName", StringComparison.OrdinalIgnoreCase))
-            {
-                result.AddInfo("Skipped 'Raven/DocumentsByEntityName' index. It is no longer needed.");
-                return;
-            }
-
-            if (string.Equals(indexDefinition.Name, "Raven/ConflictDocuments", StringComparison.OrdinalIgnoreCase))
-            {
-                result.AddInfo("Skipped 'Raven/ConflictDocuments' index. It is no longer needed.");
-                return;
-            }
-
-            if (indexDefinition.Name.StartsWith("Auto/", StringComparison.OrdinalIgnoreCase))
-            {
-                // legacy auto index
-                indexDefinition.Name = $"Legacy/{indexDefinition.Name}";
-            }
-
-            indexDefinition.DeploymentMode = IndexDeploymentMode.Parallel; //TODO - rolling index
-            var command = new PutIndexCommand(
-                indexDefinition,
-                _databaseRecord.DatabaseName,
-                null,
-                DateTime.UtcNow,
-                RaftIdGenerator.DontCareId,
-                configuration.Indexing.HistoryRevisionsNumber,
-                configuration.Indexing.StaticIndexDeploymentMode
-            );
-
-            try
-            {
-                await _server.SendToLeaderAsync(command);
-            }
-            catch (Exception e)
-            {
-                result.Indexes.ErroredCount++;
-                result.AddError($"Could not write map index '{indexDefinition.Name}': {e.Message}");
             }
         }
