+++ src/Raven.Server/Smuggler/Documents/SmugglerBase.cs
@@ -6,12 +6,15 @@ using System.Runtime.CompilerServices;
 using System.Threading;
 using System.Threading.Tasks;
 using JetBrains.Annotations;
+using Raven.Client.Documents.Indexes;
 using Raven.Client.Documents.Operations;
 using Raven.Client.Documents.Smuggler;
 using Raven.Client.ServerWide;
 using Raven.Client.Util;
 using Raven.Server.Documents;
 using Raven.Server.Documents.Expiration;
+using Raven.Server.Documents.Indexes.Auto;
+using Raven.Server.Documents.Indexes.MapReduce.Auto;
 using Raven.Server.Documents.PeriodicBackup;
 using Raven.Server.Documents.Replication;
 using Raven.Server.ServerWide.Commands;
@@ -618,7 +621,103 @@ namespace Raven.Server.Smuggler.Documents
             return result.Conflicts;
         }
 
-        protected abstract Task<SmugglerProgressBase.Counts> ProcessIndexesAsync(SmugglerResult result);
+        private async Task<SmugglerProgressBase.Counts> ProcessIndexesAsync(SmugglerResult result)
+        {
+            result.Indexes.Start();
+
+            await using (var actions = _destination.Indexes())
+            {
+                await foreach (var index in _source.GetIndexesAsync())
+                {
+                    _token.ThrowIfCancellationRequested();
+                    result.Indexes.ReadCount++;
+
+                    if (index == null)
+                    {
+                        result.Indexes.ErroredCount++;
+                        continue;
+                    }
+
+                    if (OnIndexAction != null)
+                    {
+                        OnIndexAction(index);
+                        continue;
+                    }
+
+                    switch (index.Type)
+                    {
+                        case IndexType.AutoMap:
+                            var autoMapIndexDefinition = (AutoMapIndexDefinition)index.IndexDefinition;
+
+                            try
+                            {
+                                await actions.WriteIndexAsync(autoMapIndexDefinition, IndexType.AutoMap);
+                            }
+                            catch (Exception e)
+                            {
+                                result.Indexes.ErroredCount++;
+                                result.AddError($"Could not write auto map index '{autoMapIndexDefinition.Name}': {e.Message}");
+                            }
+                            break;
+
+                        case IndexType.AutoMapReduce:
+                            var autoMapReduceIndexDefinition = (AutoMapReduceIndexDefinition)index.IndexDefinition;
+                            try
+                            {
+                                await actions.WriteIndexAsync(autoMapReduceIndexDefinition, IndexType.AutoMapReduce);
+                            }
+                            catch (Exception e)
+                            {
+                                result.Indexes.ErroredCount++;
+                                result.AddError($"Could not write auto map-reduce index '{autoMapReduceIndexDefinition.Name}': {e.Message}");
+                            }
+                            break;
+
+                        case IndexType.Map:
+                        case IndexType.MapReduce:
+                        case IndexType.JavaScriptMap:
+                        case IndexType.JavaScriptMapReduce:
+                            var indexDefinition = (IndexDefinition)index.IndexDefinition;
+                            if (string.Equals(indexDefinition.Name, "Raven/DocumentsByEntityName", StringComparison.OrdinalIgnoreCase))
+                            {
+                                result.AddInfo("Skipped 'Raven/DocumentsByEntityName' index. It is no longer needed.");
+                                continue;
+                            }
+
+                            if (string.Equals(indexDefinition.Name, "Raven/ConflictDocuments", StringComparison.OrdinalIgnoreCase))
+                            {
+                                result.AddInfo("Skipped 'Raven/ConflictDocuments' index. It is no longer needed.");
+                                continue;
+                            }
+
+                            if (indexDefinition.Name.StartsWith("Auto/", StringComparison.OrdinalIgnoreCase))
+                            {
+                                // legacy auto index
+                                indexDefinition.Name = $"Legacy/{indexDefinition.Name}";
+                            }
+
+                            await WriteIndexAsync(result, indexDefinition, actions);
+                            break;
+
+                        case IndexType.Faulty:
+                            break;
+
+                        default:
+                            throw new NotSupportedException(index.Type.ToString());
+                    }
+
+                    if (result.Indexes.ReadCount % 10 == 0)
+                    {
+                        var message = $"Read {result.Indexes.ReadCount:#,#;;0} indexes.";
+                        AddInfoToSmugglerResult(result, message);
+                    }
+                }
+            }
+
+            return result.Indexes;
+        }
+
+
 
         protected virtual Task<SmugglerProgressBase.Counts> ProcessIdentitiesAsync(SmugglerResult result, BuildVersionType buildType) =>
             ProcessIdentitiesInternalAsync(result, buildType, _destination.Identities());
@@ -1026,5 +1125,59 @@ namespace Raven.Server.Smuggler.Documents
         protected abstract Task<SmugglerProgressBase.Counts> ProcessCompareExchangeTombstonesAsync(SmugglerResult result, string databaseName);
 
         protected abstract Task<SmugglerProgressBase.Counts> ProcessCompareExchangeAsync(SmugglerResult result, string databaseName);
+
+        private async ValueTask WriteIndexAsync(SmugglerResult result, IndexDefinition indexDefinition, IIndexActions actions)
+        {
+            try
+            {
+                if (_options.RemoveAnalyzers)
+                {
+                    foreach (var indexDefinitionField in indexDefinition.Fields)
+                        indexDefinitionField.Value.Analyzer = null;
+                }
+
+                await actions.WriteIndexAsync(indexDefinition);
+            }
+            catch (Exception e)
+            {
+                var exceptionMessage = e.Message;
+                if (exceptionMessage.Contains("CS1501") && exceptionMessage.Contains("'LoadDocument'"))
+                {
+                    exceptionMessage =
+                            "LoadDocument requires a second argument which is a collection name of the loaded document" + Environment.NewLine +
+                            "For example: " + Environment.NewLine +
+                                "\tfrom doc in doc.Orders" + Environment.NewLine +
+                                "\tlet company = LoadDocument(doc.Company, \"Companies\")" + Environment.NewLine +
+                                "\tselect new {" + Environment.NewLine +
+                                    "\t\tCompanyName: company.Name" + Environment.NewLine +
+                                "\t}" + Environment.NewLine +
+                            exceptionMessage + Environment.NewLine;
+                }
+                else if (exceptionMessage.Contains("CS0103") &&
+                         (exceptionMessage.Contains("'AbstractIndexCreationTask'") ||
+                          exceptionMessage.Contains("'SpatialIndex'")))
+                {
+                    exceptionMessage = "'AbstractIndexCreationTask.SpatialGenerate' can be replaced with 'CreateSpatialField'" + Environment.NewLine +
+                                       "'SpatialIndex.Generate' can be replaced with 'CreateSpatialField'" + Environment.NewLine +
+                                       exceptionMessage + Environment.NewLine;
+                }
+                else if (exceptionMessage.Contains("CS0234") && exceptionMessage.Contains("'Abstractions'"))
+                {
+                    exceptionMessage = "'Raven.Abstractions.Linq.DynamicList' can be removed" + Environment.NewLine +
+                                       $"{exceptionMessage}" + Environment.NewLine;
+                }
+
+                result.Indexes.ErroredCount++;
+                var errorMessage = $"Could not write index '{indexDefinition.Name}', error: {exceptionMessage}" + Environment.NewLine +
+                                   $"Maps: [{Environment.NewLine}{string.Join($", {Environment.NewLine}", indexDefinition.Maps)}{Environment.NewLine}]";
+
+                if (indexDefinition.Reduce != null)
+                {
+                    errorMessage += Environment.NewLine + $"Reduce: {indexDefinition.Reduce}";
+                }
+
+                result.AddError(errorMessage);
+            }
+        }
     }
 }
